#Section 9:Authorization is completed!

+++ backend/routes/posts.js
@@ -40,8 +40,11 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
+    // console.log(req.userData);
+    // return res.status(200).json({}); // in order to prevent it to create new post before fully tested above lines
     post.save().then(createdPost => {
       res.status(201).json({
         message: "Post added successfully",
@@ -68,11 +71,17 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator:req.userData.userId }, post).then(result => {
+      console.log(result);
+      if(result.modifiedCount > 0) {
+        res.status(200).json({ message: "Post updation successful!" });
+      } else {
+        res.status(401).json({ message: "Not Authorized!" });
+      }
     });
   }
 );
@@ -114,9 +123,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator:req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if(result.deletedCount > 0) {
+      res.status(200).json({ message: "Post deletion successful!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   });
 });
